Move navigation controls into separate folder. Invert arrows.
