+++ src/components/ui/navigation-controls/navigation-controls.scss
