+++ src/components/ui/navigation-controls/navigation-cube-container.scss
