+++ src/components/ui/navigation-controls/navigation-cube-container.tsx
@@ -3,7 +3,7 @@ import React, { useRef } from "react";
 import { OrbitControls, OrthographicCamera } from "@react-three/drei";
 import { Canvas } from "@react-three/fiber";
 import { Vector3 } from "three";
-import { dstCamera } from "../../models/camera";
+import { dstCamera } from "../../../models/camera";
 import { NavigationCube } from "./navigation-cube";
 import "./navigation-cube-container.scss";
